+++ src/route2_tree_t.hpp
@@ -66,13 +66,13 @@ class route2_tree_t
         inline int _encodeNoHash(void *buf, int offset, int maxlen) const;
         inline int _getEncodedSizeNoHash() const;
         inline int _decodeNoHash(const void *buf, int offset, int maxlen);
-        inline static uint64_t _computeHash(const __lcm_hash_ptr *p);
+        inline static int64_t _computeHash(const __lcm_hash_ptr *p);
 };
 
 int route2_tree_t::encode(void *buf, int offset, int maxlen) const
 {
     int pos = 0, tlen;
-    int64_t hash = (int64_t)getHash();
+    int64_t hash = getHash();
 
     tlen = __int64_t_encode_array(buf, offset + pos, maxlen - pos, &hash, 1);
     if(tlen < 0) return tlen; else pos += tlen;
@@ -162,7 +162,7 @@ int route2_tree_t::_getEncodedSizeNoHash() const
     return enc_size;
 }
 
-uint64_t route2_tree_t::_computeHash(const __lcm_hash_ptr *p)
+int64_t route2_tree_t::_computeHash(const __lcm_hash_ptr *p)
 {
     const __lcm_hash_ptr *fp;
     for(fp = p; fp != NULL; fp = fp->parent)
@@ -170,7 +170,7 @@ uint64_t route2_tree_t::_computeHash(const __lcm_hash_ptr *p)
             return 0;
     const __lcm_hash_ptr cp = { p, (void*)route2_tree_t::getHash };
 
-    uint64_t hash = 0x429a3807c1b0e08aLL +
+    int64_t hash = 0x429a3807c1b0e08aLL +
          route2_table_t::_computeHash(&cp);
 
     return (hash<<1) + ((hash>>63)&1);
